refactor(display): split presentation builders
Move passive presentation models and helper-heavy display builders into focused modules while keeping existing import surfaces stable.

Tested: uv run ruff check src tests scripts; uv run pyright; uv run pytest -q -n 0 --tb=short tests/test_current_conditions.py tests/test_round_values_setting.py tests/test_presentation_environmental.py tests/test_weather_presenter_mobility_briefing.py tests/test_impact_summary.py tests/test_hourly_forecast_presentation.py tests/test_forecast_confidence_presentation.py tests/test_forecast_time_reference.py tests/test_openmeteo_forecast_timezone.py tests/test_show_impact_summaries_setting.py tests/test_unit_preference_bugs.py tests/test_alert_lifecycle_presentation.py tests/test_main_window_forecast_sections.py; uv run pytest -q -n 0 --tb=short

Confidence: high
