+++ src/accessiweather/display/presentation/current_conditions.py
@@ -11,7 +11,6 @@ from ...models import (
     CurrentConditions,
     EnvironmentalConditions,
     HourlyForecast,
-    HourlyForecastPeriod,
     Location,
     MinutelyPrecipitationForecast,
     TrendInsight,
@@ -22,7 +21,19 @@ from ...utils import TemperatureUnit
 from ...utils.unit_utils import format_precipitation, format_wind_speed
 from ...weather_anomaly import AnomalyCallout
 from ..priority_engine import PriorityEngine, WeatherCategory
-from ..weather_presenter import CurrentConditionsPresentation, Metric
+from .current_condition_seasonal import (
+    _build_seasonal_metrics,
+    _get_severe_risk_description,
+    format_temperature_value,
+)
+from .current_condition_trends import (
+    _adapt_temperature_trend_summary,
+    compute_pressure_trend_from_hourly,
+    describe_trend,
+    direction_descriptor,
+    format_trend_lines,
+    split_direction_descriptor,
+)
 from .environmental import AirQualityPresentation
 from .formatters import (
     format_dewpoint,
@@ -34,8 +45,21 @@ from .formatters import (
     get_temperature_precision,
     get_uv_description,
 )
+from .models import CurrentConditionsPresentation, Metric
 
 logger = logging.getLogger(__name__)
+
+__all__ = [
+    "_build_seasonal_metrics",
+    "_get_severe_risk_description",
+    "build_current_conditions",
+    "compute_pressure_trend_from_hourly",
+    "describe_trend",
+    "direction_descriptor",
+    "format_temperature_value",
+    "format_trend_lines",
+    "split_direction_descriptor",
+]
 
 
 def _normalize_metric_wind_units(value: str, *, unit_system: DisplayUnitSystem | str | None) -> str:
@@ -229,34 +253,6 @@ def _build_environmental_metrics(
     return metrics
 
 
-def _adapt_temperature_trend_summary(trend: TrendInsight, unit_pref: TemperatureUnit) -> str:
-    """Return a temperature-trend summary adapted to the user's unit preference."""
-    if trend.change is None or trend.unit not in {"°F", "°C"}:
-        return trend.summary or describe_trend(trend)
-
-    src_unit = trend.unit
-    direction = trend.direction or "steady"
-    hours = trend.timeframe_hours or 24
-
-    if src_unit == "°F":
-        change_f = trend.change
-        change_c = change_f * 5.0 / 9.0
-        if unit_pref == TemperatureUnit.CELSIUS:
-            return f"Temperature {direction} {change_c:+.1f}°C over {hours}h"
-        if unit_pref == TemperatureUnit.BOTH:
-            return f"Temperature {direction} {change_f:+.1f}°F ({change_c:+.1f}°C) over {hours}h"
-        return trend.summary or describe_trend(trend)  # FAHRENHEIT – already correct
-
-    # src_unit == "°C"
-    change_c = trend.change
-    change_f = change_c * 9.0 / 5.0
-    if unit_pref == TemperatureUnit.FAHRENHEIT:
-        return f"Temperature {direction} {change_f:+.1f}°F over {hours}h"
-    if unit_pref == TemperatureUnit.BOTH:
-        return f"Temperature {direction} {change_f:+.1f}°F ({change_c:+.1f}°C) over {hours}h"
-    return trend.summary or describe_trend(trend)  # CELSIUS – already correct
-
-
 def _build_trend_metrics(
     trends: Iterable[TrendInsight] | None,
     current: CurrentConditions,
@@ -300,137 +296,6 @@ def _build_trend_metrics(
             metrics.append(Metric("Pressure trend", legacy_pressure_trend["value"]))
 
     return metrics
-
-
-def _build_seasonal_metrics(
-    current: CurrentConditions,
-    unit_pref: TemperatureUnit,
-    precision: int,
-) -> list[Metric]:
-    """Build seasonal weather metrics (snow depth, wind chill, heat index, etc.)."""
-    metrics: list[Metric] = []
-
-    # Winter metrics
-    if current.snow_depth_in is not None and current.snow_depth_in > 0:
-        if unit_pref == TemperatureUnit.CELSIUS:
-            snow_depth_cm = current.snow_depth_cm or current.snow_depth_in * 2.54
-            metrics.append(Metric("Snow on ground", f"{snow_depth_cm:.{precision}f} cm"))
-        elif unit_pref == TemperatureUnit.FAHRENHEIT:
-            metrics.append(Metric("Snow on ground", f"{current.snow_depth_in:.{precision}f} in"))
-        else:
-            snow_depth_cm = current.snow_depth_cm or current.snow_depth_in * 2.54
-            metrics.append(
-                Metric(
-                    "Snow on ground",
-                    f"{current.snow_depth_in:.{precision}f} in ({snow_depth_cm:.{precision}f} cm)",
-                )
-            )
-
-    if current.wind_chill_f is not None:
-        # Only show wind chill if significantly different from temperature
-        temp_f = current.temperature_f
-        if temp_f is None or abs(current.wind_chill_f - temp_f) >= 3:
-            wind_chill_c = current.wind_chill_c
-            if wind_chill_c is None and current.wind_chill_f is not None:
-                wind_chill_c = (current.wind_chill_f - 32) * 5 / 9
-            wc_str = format_temperature_value(
-                current.wind_chill_f, wind_chill_c, unit_pref, precision
-            )
-            if wc_str:
-                metrics.append(Metric("Wind chill", wc_str))
-
-    if current.freezing_level_ft is not None:
-        if unit_pref == TemperatureUnit.CELSIUS:
-            freezing_m = current.freezing_level_m or current.freezing_level_ft * 0.3048
-            metrics.append(Metric("Freezing level", f"{freezing_m:.0f} m"))
-        elif unit_pref == TemperatureUnit.FAHRENHEIT:
-            metrics.append(Metric("Freezing level", f"{current.freezing_level_ft:.0f} ft"))
-        else:
-            freezing_m = current.freezing_level_m or current.freezing_level_ft * 0.3048
-            metrics.append(
-                Metric("Freezing level", f"{current.freezing_level_ft:.0f} ft ({freezing_m:.0f} m)")
-            )
-
-    # Summer metrics
-    if current.heat_index_f is not None:
-        # Only show heat index if significantly different from temperature
-        temp_f = current.temperature_f
-        if temp_f is None or abs(current.heat_index_f - temp_f) >= 3:
-            heat_index_c = current.heat_index_c
-            if heat_index_c is None and current.heat_index_f is not None:
-                heat_index_c = (current.heat_index_f - 32) * 5 / 9
-            hi_str = format_temperature_value(
-                current.heat_index_f, heat_index_c, unit_pref, precision
-            )
-            if hi_str:
-                metrics.append(Metric("Heat index", hi_str))
-
-    # Spring/Fall metrics
-    if current.frost_risk is not None and current.frost_risk.lower() != "none":
-        metrics.append(Metric("Frost risk", current.frost_risk))
-
-    # Year-round metrics - only show precipitation type when there's active precipitation
-    # Check for active precipitation indicators: precipitation amount, or condition suggests precip
-    has_active_precip = False
-    if current.precipitation_type and len(current.precipitation_type) > 0:
-        # Check if condition indicates active precipitation
-        condition_lower = (current.condition or "").lower()
-        precip_keywords = ["rain", "snow", "drizzle", "shower", "storm", "sleet", "hail", "precip"]
-        has_active_precip = any(keyword in condition_lower for keyword in precip_keywords)
-
-    if has_active_precip and current.precipitation_type:
-        precip_types = ", ".join(current.precipitation_type)
-        metrics.append(Metric("Precipitation type", precip_types.title()))
-
-    if current.severe_weather_risk is not None and current.severe_weather_risk > 0:
-        risk_level = _get_severe_risk_description(current.severe_weather_risk)
-        metrics.append(
-            Metric("Severe weather risk", f"{current.severe_weather_risk}% ({risk_level})")
-        )
-
-    return metrics
-
-
-def format_temperature_value(
-    temp_f: float | None,
-    temp_c: float | None,
-    unit_pref: TemperatureUnit,
-    precision: int,
-) -> str | None:
-    """Format a temperature value based on unit preference."""
-    if temp_f is None and temp_c is None:
-        return None
-
-    if unit_pref == TemperatureUnit.FAHRENHEIT:
-        if temp_f is not None:
-            return f"{temp_f:.{precision}f}°F"
-        return None
-    if unit_pref == TemperatureUnit.CELSIUS:
-        if temp_c is not None:
-            return f"{temp_c:.{precision}f}°C"
-        if temp_f is not None:
-            temp_c = (temp_f - 32) * 5 / 9
-            return f"{temp_c:.{precision}f}°C"
-        return None
-    # BOTH
-    if temp_f is not None:
-        if temp_c is None:
-            temp_c = (temp_f - 32) * 5 / 9
-        return f"{temp_f:.{precision}f}°F ({temp_c:.{precision}f}°C)"
-    return None
-
-
-def _get_severe_risk_description(risk: int) -> str:
-    """Get a description for severe weather risk percentage."""
-    if risk >= 80:
-        return "Extreme"
-    if risk >= 60:
-        return "High"
-    if risk >= 40:
-        return "Moderate"
-    if risk >= 20:
-        return "Low"
-    return "Minimal"
 
 
 def _categorize_metric(label: str) -> WeatherCategory:
@@ -609,142 +474,3 @@ def build_current_conditions(
         trends=trend_lines,
         impact_summary=impact,
     )
-
-
-def format_trend_lines(
-    trends: Iterable[TrendInsight] | None,
-    *,
-    current: CurrentConditions | None = None,
-    hourly_forecast: HourlyForecast | None = None,
-    include_pressure: bool = True,
-    unit_pref: TemperatureUnit = TemperatureUnit.FAHRENHEIT,
-) -> list[str]:
-    """Generate human readable summaries for trend insights."""
-    lines: list[str] = []
-    pressure_present = False
-
-    if trends:
-        for trend in trends:
-            metric_name = getattr(trend, "metric", "")
-            if isinstance(metric_name, str) and metric_name.lower() == "daily_trend":
-                continue
-            is_pressure = isinstance(metric_name, str) and metric_name.lower() == "pressure"
-            if is_pressure and not include_pressure:
-                continue
-
-            is_temperature = isinstance(metric_name, str) and metric_name.lower() == "temperature"
-            if is_temperature:
-                summary = _adapt_temperature_trend_summary(trend, unit_pref)
-            else:
-                summary = trend.summary or describe_trend(trend)
-
-            if trend.sparkline:
-                summary = f"{summary} {trend.sparkline}".strip()
-            if summary:
-                lines.append(summary)
-            if is_pressure:
-                pressure_present = True
-
-    if include_pressure and (not pressure_present) and current and hourly_forecast:
-        legacy_pressure = compute_pressure_trend_from_hourly(current, hourly_forecast)
-        if legacy_pressure:
-            summary = legacy_pressure["summary"]
-            if summary and summary not in lines:
-                lines.append(summary)
-
-    return lines
-
-
-def describe_trend(trend: TrendInsight) -> str:
-    """Describe a trend insight when it lacks a canned summary."""
-    direction = (trend.direction or "steady").capitalize()
-    timeframe = trend.timeframe_hours or 24
-    change_text = ""
-    if trend.change is not None:
-        if trend.unit in {"°F", "°C"}:
-            change_text = f"{trend.change:+.1f}{trend.unit}"
-        elif trend.unit:
-            change_text = f"{trend.change:+.2f}{trend.unit}"
-        else:
-            change_text = f"{trend.change:+.1f}"
-    pieces = [direction]
-    if change_text:
-        pieces.append(change_text)
-    pieces.append(f"over {timeframe}h")
-    return " ".join(pieces)
-
-
-def compute_pressure_trend_from_hourly(
-    current: CurrentConditions,
-    hourly_forecast: HourlyForecast | None,
-) -> dict[str, str] | None:
-    """Approximate a pressure trend using upcoming hourly data."""
-    if not hourly_forecast or not hourly_forecast.has_data():
-        return None
-
-    next_hours = hourly_forecast.get_next_hours(6)
-    if not next_hours:
-        return None
-
-    target: HourlyForecastPeriod | None = None
-    for period in reversed(next_hours):
-        if not period:
-            continue
-        if period.pressure_in is not None or period.pressure_mb is not None:
-            target = period
-            break
-
-    if target is None:
-        return None
-
-    base_in = current.pressure_in
-    base_mb = current.pressure_mb
-    future_in = target.pressure_in
-    future_mb = target.pressure_mb
-
-    descriptor: str | None = None
-    magnitude: str | None = None
-    change: float | None = None
-
-    if base_in is not None and future_in is not None:
-        change = future_in - base_in
-        descriptor = direction_descriptor(change, minor=0.02, strong=0.05)
-        magnitude = f"{change:+.2f} inHg"
-    elif base_mb is not None and future_mb is not None:
-        change = future_mb - base_mb
-        descriptor = direction_descriptor(change, minor=0.5, strong=1.5)
-        magnitude = f"{change:+.1f} mb"
-
-    if descriptor is None or magnitude is None or change is None:
-        return None
-
-    direction_word, arrow = split_direction_descriptor(descriptor)
-    window_hours = 6
-    arrow_part = f" {arrow}" if arrow else ""
-    value = f"{direction_word.title()}{arrow_part} {magnitude} over next {window_hours}h".strip()
-    summary = f"Pressure {descriptor} {magnitude} over next {window_hours}h".strip()
-
-    return {"summary": summary, "value": value}
-
-
-def direction_descriptor(change: float, *, minor: float, strong: float) -> str:
-    """Classify change magnitude into rising/falling/steady buckets with arrows."""
-    if change >= strong:
-        return "rising ⬆⬆"
-    if change >= minor:
-        return "rising ⬆"
-    if change <= -strong:
-        return "falling ⬇⬇"
-    if change <= -minor:
-        return "falling ⬇"
-    return "steady →"
-
-
-def split_direction_descriptor(descriptor: str) -> tuple[str, str]:
-    """Return the textual direction and accompanying arrow glyphs."""
-    if not descriptor:
-        return "steady", ""
-    if " " in descriptor:
-        direction, arrow = descriptor.split(" ", 1)
-        return direction, arrow.strip()
-    return descriptor, ""
